+++ examples/localization/src/main.rs
@@ -3,10 +3,10 @@
 use bevy::prelude::*;
 use bevy::window::WindowTheme;
 use bevy_cobweb::prelude::*;
+use bevy_cobweb_ui::builtin::widgets::radio_button::{RadioButtonBuilder, RadioButtonManager};
 use bevy_cobweb_ui::prelude::*;
 use bevy_cobweb_ui::sickle::ui_builder::*;
 use bevy_cobweb_ui::sickle::SickleUiPlugin;
-use bevy_cobweb_ui::widgets::radio_button::{RadioButtonBuilder, RadioButtonManager};
 
 //-------------------------------------------------------------------------------------------------------------------