+++ src/builtin/colors/mod.rs
@@ -1,0 +1,3 @@
+mod plugin;
+
+pub(crate) use plugin::*;